refactor: relocate "PCMsg" and related types to TUI
