+++ tui/src/ui/components/podcast.rs
@@ -8,7 +8,6 @@ use serde_json::Value;
 use termusiclib::config::SharedTuiSettings;
 use termusiclib::ids::Id;
 use termusiclib::podcast::{EpData, PodcastFeed, PodcastNoId, download_list};
-use termusiclib::types::PCMsg;
 use tokio::runtime::Handle;
 use tui_realm_stdlib::List;
 use tuirealm::command::{Cmd, CmdResult, Direction, Position};
@@ -21,7 +20,7 @@ use tuirealm::{
 
 use crate::ui::Model;
 use crate::ui::model::UserEvent;
-use crate::ui::msg::{GSMsg, Msg};
+use crate::ui::msg::{GSMsg, Msg, PCMsg};
 
 #[derive(MockComponent)]
 pub struct FeedsList {
@@ -444,7 +443,7 @@ impl Model {
             ),
             &self.taskpool,
             move |msg| {
-                let _ = tx_to_main.send(Msg::Podcast(msg));
+                let _ = tx_to_main.send(Msg::Podcast(PCMsg::SyncResult(msg)));
             },
         );
     }
@@ -664,7 +663,7 @@ impl Model {
                 ),
                 &self.taskpool,
                 move |msg| {
-                    let _ = tx_to_main.send(Msg::Podcast(msg));
+                    let _ = tx_to_main.send(Msg::Podcast(PCMsg::SyncResult(msg)));
                 },
             );
         }
@@ -766,7 +765,7 @@ impl Model {
                         ),
                         &self.taskpool,
                         move |msg| {
-                            let _ = tx_to_main.send(Msg::Podcast(msg));
+                            let _ = tx_to_main.send(Msg::Podcast(PCMsg::DLResult(msg)));
                         },
                     );
                 }
